requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.11.9
pydantic-settings==2.1.0
openai==1.97.0
sentence-transformers==5.1.2
faiss-cpu==1.13.1
rank-bm25==0.2.2
numpy==2.2.6
python-dotenv==1.0.0
streamlit==1.37.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
aiofiles==24.1.0
